removing unnecessary file
adding SU checks to another

+++ host/bin/stop.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
 
+if [[ $UID != 0 ]]; then
+    echo "Please run this script with sudo:"
+    echo "sudo $0 $*"
+    exit 1
+fi
+
 pkill -f org.area515.resinprinter.server.Main
